Added debug code to frame_manager
Now can print out a map of the frame allocation, splitted ranges of
allocated/unallocated frames. Also added a test for it in kernel_tests.
Finally, added atoi_base_u, which operates on unsigned numbers, and have
puth redirect to it. This allows me to not write negative hexadecimal
numers, which are used generally for adresses